OTWO-4790 - added script to migration
